Add tests, remove unused helper
